[VisionGlass] Fix controller orientation issues
Controller orientation was suffering from gimbal lock. This means
that rotations higher than 90º started to exhibit weird behaviours
like inverted or swapped axis. We tried to fix that in the past
without much success, but now we do now the actual culprit of that
gimbal lock and can fix that once for all.

The quaternion returned by the phone sensor is using Android axis.
If the phone is on a table in horizontal position and portrait
orientation with respect to the user, X goes right, Y goes forward
and Z goes up. However in our world coordinates, X goes right (same)
but Y goes up and Z goes backward (-Z is forward direction). So we
have to transform the quaternion to follow that coordinate system.

Once that was fixed gimbal lock dissappeared. The only missing thing
was to fix the calibration code removing some workarounds we had
to add to get it "working" with the previous incorrect version of
the orientation code.

Last but not least, we changed the sensor delay from NORMAL to
GAME to get a little bit more accuracy in tracking.

As nice side effect this also fixes orientation issues when
entering WebXR experiences. Fixes #1522