Use a different Android property to get the OS version
Recent upgrades in the Pico OS changed the way OS releases are
named. We depend on those release numbers to enable/disable
some workarounds we had to add for runtime issues.

What was not changed is the property that stores the OS version
id to be displayed in the UI, which is ro.build.display.id so
we are going to use it from now on.

Apart from that we're adding an extra security layer by adding
a fallback in case the version number cannot be parsed. In that
case we will always assume that we have the most recent version
of the OS and won't apply any workarounds. This fixes a crash
when launching Wolvic on the new Pico4 Ultra.