[Oculus] Fix immersive environments
To construct the immersive environment, Wolvic creates an OpenXR
cubemap layer and applies textures to it.

For Oculus, we were using compressed textures loaded from KTX files.

Starting with version 69 of the Meta OS, this method stopped working
and the environments just looked pitch black. It seems that the
problem is caused by a bug on the system that breaks the use
of compressed textures on a cubemap layer.

While this bug is fixed upstream, this patch forces Oculus devices
to load the immersive environments from PNG files.

Fixes https://github.com/Igalia/wolvic/issues/1535