Compare semantic versions
Add a method StringUtils.compareVersions() that compares semantic
versions. We assume that versions are defined as integer numbers separated by periods ('.'), but the method also takes into account
empty strings and the possibility that some components can not be
parsed as integers.

Use StringUtils.compareVersions() to check the version of the
PICO operating system when enabling passthrough.

Tests are included.