Initial implementation of browser tabs
The Tabs Bar is a small window that sits next to the current window
and which lists the open tabs.

There are two implementations:
  - HorizontalTabsBar on the top of the window
  - VerticalTabsBar on the side of the window

The Tabs Bar is managed by VRBrowserActivity.

SessionStore now keeps a list of session change listeners.

TabDelegate is moved to a separate interface.

The current tabs style can be changed in Settings -> Display.

WindowViewModel gets three new fields:
- isTabsBarVisible
- usesHorizontalTabsBar
- usesVerticalTabsBar

TopBarWidget uses two of these fields to update its placement
so as to not overlap with the tabs bar.

Note that we don't (yet) link windows and tabs, so the interface
gets a bit confusing when we have more than one window open.

WidgetPlacement gets two new fields: horizontalOffset, verticalOffset.
These define a translation offset (in world dimensions), which is
applied after the anchors have been calculated. The goal is to be able
to move widgets from their initial position in a predictable way, to
make space for the tabs bar when necessary.