[AOSP] Migrate to GameActivity
NativeActivity was the original choice for the Android activity for
the AOSP port because it's a natural choice for an activity with
native code. That was causing issues in MagicLeap2 because key
strokes in external bluetooth keyboards were filtered out.

By switching to GameActivity we can get rid of that limitation.
The migration only required an small adjustment in the back button
handling code because the onKeyUp() event for the back button was
not triggering the onBackPressed() as it does in NativeActivity.