[PicoXR] Use compressed ETC2 textures for environments
PicoOS v5.4 added support for ETC2 compressed textures. We removed
support for PicoOS < v5.7 when the version check was added. This means
that we can now finally switch to using compressed textures also
for Pico devices.

There is no need to distribute the PNG versions in the Pico package.
That would save us 14MB in the distributed APK. Apart from that
the external environments code was also modified to download the
KTX versions of the textures.

The already downloaded environments will be used though because
we have the same id for the different formats of the same environment.
But for the builtin ones, and new downloads the KTX versions
would be the selected choice.

Fixes #522