[Chromium] Show private mode about page
When entering private mode Wolvic displays a custom web page
that explains the user what private mode is and what is not.

In the case of Chromium backend it was not showing anything
because the Session::loadData interface was not implemented.
This PR implements it using the new API added in
https://github.com/Igalia/wolvic-chromium/pull/119