Optimize UrlUtils::isPrivateAboutPage
The isPrivateAboutPage was called 18! times just during the startup
process of Wolvic. That method was doing a couple of things very badly:
1. Most of the time users would load http(s) URLs and this method was not
filtering out non-data:// URLs at the very beginning
2. For each call (as long as URL was not null) it was creating the private
about page and encoding it in base64.

We can do it much better by:
1. filter out non-data:// URIs
2. create and decode the about page on demand just once as it never changes

With this optimization we moved from 18 about page creations to 0
(when using the default start page).