Add support for MSFT hand interaction profile
Support for the Khronos XR_EXT_HAND_INTERACTION extension was added
not so long ago. That extension defines a new interaction profile
that can be used to get the input from hands using the same OpenXR
concepts that are valid for physical controllers.

There is another vendor extension that does the same which is
XR_MSFT_HAND_INTERACTION which is supported in devices like Meta
devices (Quest2, Quest3, QuestPro) or Lynx-R1.

Even if they are 2 different extensions, the nice thing is that we
could use the same code path for both reducing the mess and complexity
of the hand tracking code. Not only that, as they both define
interaction profiles, they integrate nicely with the code that
handles physical controllers.

Although unlikely, it's possible that a given runtime provides both
the Khronos and MSFT extensions. In that case we select the
Khronos one.