Improve system gesture detection with hand interaction profiles
Some actions are linked to specific hand gestures. By facing the
palm of each hand and pinching one could go back (left hand) or
exit Wolvic (right hand).

When using hand interaction profiles (XR_EXT_hand_interaction,
XR_MSFT_hand_interaction) we read the state of the actions
that are bound to the paths provided by the profiles. In our
case we get both trigger events (pinches) and squeeze events
(closing fist).

The key here is that those events are not mutually exclusive,
actually when pinching we're partially closing the fist. That
means that whend doing the pinch action we could receive
a 0.9 value for pinching but also a 0.5 value for squeezing.

The code was not discerning both cases properly when the hand
action was detected (palm facing head). That lead to situations
in which those pinches were not properly detected because when
iterating over the actions we were
1. detecting a pinch -> setting APP button as clicked
2. detecting a partial squeeze (not passing the threshold)
   -> setting APP button as released

That sequence was effectively not allowing pinch events to
be detected because the click status was cleared before
upper layers could detect those back/exit actions. The
consequence was that those hand gestures were not properly
detected. We had a workaround in place, which basically
was lowering the threshold but that was just a hack
that was not addressing the actual problem.