Remove overscroll mode
Set the "overScrollMode" to "never" by default and remove those instances
where it had been manually set.

The overscroll mode is the responsible for the bouncing and rubber-band
effect that happens when we reach the end of a scrollable list. This
effect makes sense in a touchscreen but it looks jarring on VR.