Do not filter out button state processing when ready action is false
The XR_EXT_hand_interaction defines several "ready" actions for some of
their paths to inform the clients of the API that the tracking system
considers that the hand shape is observed ready to perform some
gesture (pinch, grasp...).

We were incorrectly ignoring the "value" of the associated path in
those cases. However that is wrong for some reasons:
1. "ready" is only true when the gesture is almost completed, this
means that we'd be ignoring all the values before the gesture is
detected. For example the "value" for pinch returns a float between
0.0 and 1.0 indicating how close to a pinch a given gesture is.
2. if "ready" is false, we were not updating the value of the
button, meaning that a trigger button will remain always in the
clicked state after the first pinch happened.

This was detected in Meta's OS v71 beta version.