Remove AppLab build
Meta discontinued AppLab as devs can now publish apps in the Meta Store
in beta status for people to try, which is more or less what AppLab was
for.

We can get rid of the build options for AppLab and also the deprecation
dialog we were showing in AppLab builds encouraging users to switch to
the Meta Store version.