Improve scrolling using eye tracking
This PR fixes a couple of issues with the current gesture to scroll
using eye tracking:
1. it allows users to scroll with both hands, initially it was only
tied to the right hand
2. it prevents scrolling by moving the eyes. The scroll gesture
should be done with the hand or the controller but not with the gaze.
The current code scrolls if user moves their eyes up and down while
pinching/clicking and holding.

To fix that we pass a new parameter to the Update() method in the
input source carrying the transform for the eye gaze. That transform
would be used to set the starting point of the scroll action if
there is a scroll or the pointer position otherwise.

Methods like Update(), EmulateControllerFromHands() or
HandleEyeTrackingScroll() are suffering from bad design
decisions and this PR does not improve the situation. In any case
that's a different topic and should be addressed in a separate PR.

Tested in:
* Pico 4E (5.11.1)
* Meta Quest Pro (v71)
* Magic Leap 2