Launch immersive experiences directly
This PR implements support for opening immersive experiences
directly as soon as the application is launched.

We do this by adding a built-in extension which will activate
a particular element in the page.

We need to set the preference dom.vr.require-gesture to false
so this script can launch immersive WebXR experiences.

Media autoplay is also allowed in this mode.

The information required to identify the element to activate
is passed as parameters to the Intent:

- launch_immersive
- launch_immersive_parent_xpath
- launch_immersive_element_xpath
- a target URL to open

One additional concern is that we should only change the
configuration prefs when that is needed for launching
Wolvic directly in immersive mode.

The problem is that we only write that prefs file the first time
that we create the browser engine, which happens several times at
launch time **before** we have parsed the incoming Intent and we
can know in which mode we are launching.

The solution is that WidgetManagerDelegate.isOpenInImmersive()
internally will check the current Intent, to cover the case
where it is being called too early.

We then pass VRBrowserActivity as the context to the objects
that may initialize the engine, so eventually
SessionUtils.prepareConfigurationPath() is able to decide whether
specific prefs should be added to the configuration file.

For launching Wolvic directly in immersive mode, these prefs are:
  dom.vr.require-gesture: false
  media.autoplay.default: 0

These preferences need to be applied on startup. Therefore, when
a running Wolvic receives a new Intent to launch in immersive we
will finish the current activity and launch a new one.