Add a move bar widget
This widget replaces the experimental move windows button that could
be enabled from the Display Settings. That button was shown in the
navigation bar and was a bit weird to use.

This PR replaces it with a thing semi-transparent bar that when
dragged moves the window in the horizontal direction as the old
button used to do. That bar is unconditionally shown when in
curved mode.