Sync UI with pinch gesture detection
When using hand tracking in some systems the pointer was not getting
azure (signal of pinching) until the fingers were actually in contact.
However for the pinch to happen, there is no need to touch them,
once they are close enough we consider them performing the pinch
gesture.

The problem was that the code detecting the gesture and the code
changing the color of the pointer were not using the same activation
thresholds. Actually the color of the pointer was only changed
when the pinch gesture was returning 1.0, i.e, fingers touching.