Do not hide the hands with no aim and no gesture
There was an early return disabling the hand controller whenever
we had no aim and no system gesture was detected. The rationale
was that if we have no aim then we must be doing a system gesture.

However that's a false assumption. When the hand is in the process
of rotating there is an angle in which there is no aim but also
there is no system gesture (the hand is not facing the head).
Removing that condition makes the hand not to vanish.

Apart from that we have verified that Pico fixed the bug in the
runtime that was preventing the FB aim extension from correctly
detecting system gestures. We can safely use it now.

Fixes #1584