Fallback code for unknown Meta and Pico devices
Add fallback logic for the case where  we fail to recognize the model
of the current device.

If the manufacturer is "Pico" and the device has 6DoF we will assume
that it is compatible with the PICO4.

If the manufacturer is "Oculus", the device has 6DoF and the model
name begins with "Quest",  we will assume that it is compatible with
the Meta Quest 3.