Reimplement FB passthrough layer (#1607)
The current code enables FB passthrough and the passthrough layer
on start if supported by the runtime. Instead of doing that, we can
create both the passthrough handler and the passthrough layer and
activate them on deman. This will be done for WebXR AR experiences
and also when activating passthrough in the 3 dots menu.

Apart from that we've removed a lot of the boilerplate that was
not really needed. The code used to threat the passthrough layer
as any other compositor layer that had to be integrated into the
VRB scenegraph. The passthrough layer is a compositor layer that
has nothing to do with Wolvic's scenegraph, and thus, does not
need to be wrapped by a VRLayer nor added as a node.

There should not be any visual changes from the user POV. However
it would allow us to to have things like issue #1351 in the future.