[OpenXR] Remove DoF filtering
We were filtering by DoF in order not to suggest some profiles for
devices not supporting 6DoF controllers. On the one hand, this was
only needed for HVR so we better handle it as an exception. And on
the other hand, there is really no need to filter out profiles in
general as the runtime will select the most appropiate one.