Use the widget name from widget placement
The WidgetPlacement structure contains the name of the widget. When
building the name for the native widgets we were using a combination
of "Widget-" prefix + the widget handle. That was not useful at all
for debugging as it was not possible to differentiate between widgets
by name.

Instead it's much better to use the widget name that is passed in
WidgetPlacement. That way we can identify whether a widget is a
TrayWidget, a WindowWidget... We keep the widget handle in the name
as it's useful to differentiate widgets that may have multiple
instances (like Windows for example).