Indicate content type with getExternalFilesDir()
We used getExternalFilesDir() to get the location where environments
and dictionaries will be downloaded. It is better to pass a content
type to this call so the files are not saved in the root folder.

In particular, using DIRECTORY_DOWNLOADS makes it easier to avoid
including these files in a backup.