Revert "Synthesize FCP for cached pages"
This reverts commit aae801484e40c68b9c4e1989b29ea2e3c8059476.

This workaround was added after the update to Gecko 121. The issues
with first contentful paint are not here anymore so we can remove
it and avoid synthesizing FCP events at the wrong time which were
causing issues when taking screenshots of the web contents.

When those generated FCPs were issued, the Gecko compositor was
not in a ready state to generate a screenshot and thus an exception
was generated. Gecko was not handling it properly so it was
affecting subsequent JNI calls causing crashes and hangs in Gecko.

Fixes #1638