Emit onFirstContentfulPaint paint for cached pages
The FCP event is not emitted by GeckoView if the page is loaded from the cache
(either HTTP or back/forward). As a result many things are not properly set up
in Wolvic, like the current active session, resulting in many cases in web
contents not visible on app start.

In those cases we should generate and emit it to get things working.