Check storage permissions before installing local addons
Addons can be installed from locally downloadable XPI files with the correct
MIME type. Installing those local files require android storage permissions.
Those permissions were not checked when installing (and are not granted by
default) so local installations used to silently fail in those cases.

From now on a dialog is shown in case the permission is not granted.