Better tracking of the specific window content type
Replace the integer constants for the library sections in WindowWidget
with an enum that also includes the about: URL for each one.

WindowViewModel.currentContentType keeps track of the specific content
being shown in the window at each moment, which will make it easier to
implement the native UI New Tab page and reorganise the library.

WindowViewModel.isNativeContentVisible depends on the previous field
and provides a boolean value so the current bindings keep working.

This PR should not cause any major behaviour changes.