Refactor head lock mode to enable controller lock modes
When in head lock mode, the view is automatically reoriented to where the
head is looking at. In this mode the window is always in front of user's
head. We could use the same approach to implement moving windows
behaviour. The first step is to generalize the lock mode to also
support other modes apart from head lock mode.