Move windows in the Z direction using the moving bar widget
By moving the controller closer or away users can now freely position
the windows also in the Z axis. So far it still respects the current
limits for near and far window distances.

This allows us to remove the slider to change the window position
in the settings as it is no longer needed.