Migration to Gecko 128.5.1esr
The recent backports to the 128.x ESR releases brought changes
to the shared memory structures we use for WebXR. Most of them
are migrations from C arrays to std::arrays but there are some
other changes.

Apart from that the structures used in Chromium for hand tracking
and AR had to be updated too.