Add a link to Add-ons in the Settings
Add a button to open the Add-ons page from the Settings, since that
is a pretty obvious place to find this functionality.

To avoid messing up the layout of the honeycomb buttons, I've moved
the "What's New" button closer to the version details. After some
experimentation, I've placed two small buttons on the right side of
the Settings window, linking to the Feedback and News pages.

The background of the What's New button becomes highlighted when the
details of the new version are available (this replaces the little
blue dot that we had before).

I have also simplified the implementation of the hidden shortcut to
show the version details. Now it is just a matter of clicking on the
app logo or the app name.

Finally, I have also simplified the layout in the Settings window by
removing an unneeded LinearLayout and using the parent
ConstraintLayout instead.