Replace GitHub link with regular feedback page
Use the regular feedback page in the "Feedback" link in the Settings,
instead of taking users to GitHub. This is the same link that is used
in a similar functionality in the navigation bar.

It is unexpected that two similar "feedback" buttons would take the
user to two different pages. Furthermore, most users do not have a
GitHub account.