Prevent a crash in the slider widget
The Slider widget will throw an exception if we try to set a value
which does not respect the current step size, and min and max values.

This can happen accidentally, for example when restoing the app's
state from disk.

To prevent similar crashes in the future, this change ensures that
the value is acceptable, and rounds and clamps it if necessary.