Size and spacing of navigation bar elements
Set the default size in dp (800dp, previously 720dp) of the
navigation bar to be consistent with its default size in world
dimensions (4.0).

This makes the navigation elements slightly smaller, which we
compensate by increasing their size to 44dp.

Reduce the margin between elements.

Split the styles for the navigation bar from the ones used in other
elements in the UI, for example in the close button in dialogs.