Do not show builtin extensions (addons) in the hamburger menu
Wolvic uses builtin extensions (not user installable) for several matters,
from better YouTube experiences to workarounding issues in media playing.
Those builtin addons are not user installable and they can't be disabled
or uninstalled either. Actually the point is to make them completely
transparent to the users. That's why we should not show them in the
hamburger menu.