Use the move bar widget to freely position the windows
Users should be able to move the windows around (although the 3 windows will still be
grouped together) by clicking and dragging in the recently added move bar bellow
the browser front window.

This required a small refactoring in head lock mode. From now on, instead of having
a head lock on/off switch, we have a head lock mode, which can take three different
values: no lock (normal mode), head lock and controller lock (the one used to move
windows).

Currently when moving windows, they closely follow the orientation of
the controller. That means that they are prone to the natural small
shaking of human hands. To alleviate that, we can add some linear
interpolation between consecutive orientations. By using a value close
to 0 (in this case 0.1) we ensure that the interpolated value is very
close to the previous value, and thus, the movement is very smooth.

Fixes #99, fixes #927