Enable smooth window movement
Currently when moving windows, they closely follow the orientation of
the controller. That means that they are prone to the natural small
shaking of human hands. To alleviate that, we can add some linear
interpolation between consecutive orientations. By using a value close
to 0 (in this case 0.1) we ensure that the interpolated value is very
close to the previous value, and thus, the movement is very smooth.