State of tray buttons for Bookmarks and Downloads
Ensure that the state of the tray buttons for bookmarks and downloads
stays consistent in all cases.

This fixes several bugs, including one where the state of these
buttons was not updated when the user navigated to another section
within the Library, and another where the icon padding was not always
updated correctly.

The main change is that LibraryPanel defines a Controller interface
that is used to set the current panel content. This is called from
the navigation buttons within the Library, so the containing
WindowWidget is able to update its state accordingly.