[OpenTelemetry] Inital local only implementation
This is an initial implementation of the ITelemetry interface
using OpenTelemetry, in particular, the opentelemetry-android
library.

Using this new library forces us to bump the required kotlin
version and the project compilesdk version.

So far the telemetry data is logged only to the logcat and not
sent to any server. In order to avoid too many writes (and in
the future to batch requests to the servers) it uses on disk
buffering.