[VisionGlass] Use the proper registerReceiver() call depending on Android version
Starting on API level 34 (UpsideDownCake) either RECEIVER_EXPORTED or
RECEIVER_NOT_EXPORTED must be specified if the receiver is not being
registered for system broadcasts or a SecurityException will be thrown.

We were unconditionally doing that, however the min API version for
the VisionGlass is API level 24, which does not have neither the new
registerReceiver() call with 3 arguments nor the RECEIVER_* flags.