Reduce sensitivity required by controller trigger
We've been traditionally using a high value for clicking threshold
(0.91) in order not to generate false positives specially when using
hand tracking. However, that does not make much sense when using
physical controllers because they're very precise and false positives
are not a real problem there.

We can actually lower it to just 25% of the full pressed position
without any accuracy loss, and indeed, the experience is much
better. It's also good for a11y.

Fixes #1487