[PicoXR] Do not use XR_FB_hand_tracking_aim
This vendor extension provides very good values for aim when using
hand tracking and is also able to detect some common gestures
like pinching or doing a system gesture. It has never been really
advertised by the runtime as supported but we knew it worked so
we were using it anyway.

However the behaviour in the latest versions of the OpenXR runtime
from Pico has degraded a lot, to the point that is no longer able
to detect those hand gestures. This means that we could only use
the aim at the most. Provided our computed aim is good enough
it's better to completely disable the extension for Pico.

A nice side effect, is that the ray now properly originates from
the hand. With the extension the ray origin was shifted upwards.

Fixes #1679