[Lynx] Workaround a runtime bug not reporting tracked hand positions
Lynx runtime never returns the XR_SPACE_LOCATION_POSITION_TRACKED_BIT
flag as active, even if it's really actively tracking the position
of the hands. That confused the code and made it think that emulated
positions were returned instead.

That was causing the origin of the aim pose (and thus the origin
of the aim ray) to be incorrectly shifted from the hands. The
workaround is replacing the check for POSITION_TRACKED_BIT by
POSITION_VALID_BIT. That way we'd still be able to detected when
hands are not tracked while providing a much better aim.