Hack for the initial rotation
When moving windows with the new bar widget the windows initally move down
as soon as the user grabs the bar. That's because the reorientation operates
over the front window and not over the new bar widget.

In order to make it slightly better we're compensating part of that initial
rotation. But this is an ugly hack for the 1.8 release that should be
removed as soon as possible.