Apply border and brightness effects
This patch applies border and brightness effects to OverlayContentWidget.

The border will overlap 2px on the left/right/top/bottom of
WebContents because the browser's compositor implementation doesn't
consider the x, y coordinates when setting up the surface. It's same
with Widget for Tab, and we maybe consider to have the inner Widget
to avoid this issue later. Having a border effect will bring about
a better UX even if it loses 2 pixels.