[VisionGlass] Use TYPE_ROTATION_VECTOR instead of TYPE_GAME_ROTATION_VECTOR
The GAME rotation vector does not use the geomagnetic field, therefore the Y
axis does not point north. But according to the docs,
"However, the orientation may drift somewhat over time". That's why we
switch to ROTATION_VECTOR to avoid that drift.

As the new type of rotation vector points to the north, there is no need
to perform re-initializations every time the controller is calibrated as
the origin of coordinates won't change.