Fix undetected click events on some OpenXR runtimes
Wolvic has had since very early on, a code that considered button
presses over a threshold as clicks (full presses) because some
old devices only reported clicks with a 100% pressed value
which was not very reliable. This meant that values > 91% were
considered clicks as well.

We recently landed a patch that lowered that limit for physical
controllers because they're much more reliable now and this way
users didn't have to fully press them which is better for a11y.

That change unveiled a bug that has been there for a while.
Runtimes don't often return the clicked bit set unless the
pressed value is 100%. However some others (like the ML2 one)
do. The condition that was setting the clicked value to true
if the value was above a certain limit assumed that the runtime
was not setting the clicked value to true. That's true for most
of runtimes but not MagicLeap's which works better.

Fixes #1703