[Chromium] Do not show addons icon in Settings dialog
The same way we don't show them in the hamburger menu there is no
point in showing it if there is no support for extensions yet.

Changing this mad the "what's new" and feedback buttons to clash.
I modified the layout so that they're both aligned with the start
and end of the settings dialog respectively. This way they won't
interfere with each other in any case.

Fixes #1713