Add a new Sync button in the tabs bar
In order to receive tabs through the FxA SendTabTo service we need
to poll the API; the new button will perform this action.

Ideally, we should be able to observe events from the FxA push
service, but for now we need to depend on user intervention to
ask the server for new tabs.