[PicoXR] Remove version check to enable passthrough
Passthrough was enabled in PicoOS 5.7.1 via a FB OpenXR extension.
We added a version check to avoid enabling passthrough in unsupported
OS versions. That version number was retrieved from a system property,
and lexicographically checked against "5.7.1". The problem is that
"5.10" in indeed < "5.7" because we're not comparing the versions but
the whole string.

The solution is to remove the check because it's pointless now as
it's a very old version. We actually removed those checks in native
code long time ago.

Fixes #1720