Synthesize a move up event when a controller is disabled while dragging
When doing a click and hold over a widget (like when scrolling) it might
happen that the controller becomes unavailable, for example, if it
goes out of sight. In those cases we must stop the dragging because
otherwise the widget will think that the drag is still taking place
and will refuse to accept any other event (like clicks with other
controllers).

Fixes #1712