Permission switches in Settings remain consistent
The state of permissions in the Privacy Settings can sometimes be
inconsistent. This affects specially the location permissions,
because the user is able to grant us access to either the "coarse"
or the "fine" location using the system settings.

This change ensures that the UI is updated consistently both when
we create this dialog and when a new permission is granted.