[OpenTelemetry] Setup dependencies and required sw versions
This is a preparation commit for the integration of the
opentelemetry-android library into Wolvic that will allow us
to know more about users actions and their usage of the
different features (obviously with their prior consent).

The OpenTelemetry android library requires at least a compilesdk 35
and kotlin 2.1 so let's bump them both. Also we need to enable
desugaring for the library to work properly with newer versions
of Android.

We use a different approach to include new dependencies. Instead
of using the custom versions.gradle file, we switched to the
more modern library catalog declarations.