Open tabs in background by default
This change restores the behaviour of opening new tabs in the
background by default.

This functionality had been changed when the new tabs bar was
implemented. The reason was that opening a tab in the background
leaves it in an uninitialized state until the user clicks on
that tab, and this made it difficult to keep the tabs bar
updated in real time.

We will use the domain name as the title of background tabs
until the user clicks on them (this was empty previously).