[PicoXR] Use a different input profile for Pico4U
Pico has defined a input profile for the Pico4U which is different
to the ones for the Pico4 or the Pico4E. It is not in any OpenXR
doc so we were not aware of it. We got that info from the Pico
customer service.

The problem was that the runtime was selecting the generic Khronos
profile instead of the Pico4 profile because we were not suggesting
the specific profile for the 4U but for the 4 (or 4E). Actually
the runtime should probably select the Pico4 profile over the generic
one but that's their decision anyway.