[PicoXR] Remove the Pico4xOld interaction profile
That interaction profile was used long time ago when the Pico4 device
didn't have its own interaction profile. It's actually the profile
designed for the Neo3. Recenctly Pico released an extension that
provide a specific profile for the Pico4, so this should not be
needed. Actually the code was not using it at all, so any device
selecting it would be broken.