[PICO] Disable the controller when no position info is available
PICO continuously tracks the controllers even when they are not
in use, so we always check for additional hand tracking
information.

However, it might happen that the PICO controller has the
SPACE_LOCATION_ORIENTATION_VALID flag unset and at the same
time we are not able to get hand tracking information.

In that scenario, we should hide the controller.