Apply MediaSession fixes to all videos
Wolvic has an internal extension which feeds the navigator.mediaSession
object with progress information. It was initially developed to get
progress information for the progress bar that is shown on fullscreen
videos. That's the reason why the extension was only injecting data
into the web engine for those videos that were emitting the onfullscreen
events.

The problem is that in VisionGlass we have another progress bar that
is shown in the phone UI but for all videos, not only fullscreen ones.

The fix is to capture the playing event and then apply the mediasession
fixes that we already have in place for the video element that
issued the event.

Fixes #1406