Remove the WaveVR backend
This came from FxR and was mainly used by HTC devices. We haven't
released any official version of Wolvic for HTC so this means that
this port has been basically unmaintained for years. It has been
mainly a maintenance burden forcing us to upgrade APIs just for
the sake of keeping it building. It's time to remove it.