[VisionGlass] Avoid asking for USB permission when connecting the device
If Wolvic is defined as the default application for a USB, it's possible
to avoid requesting for USB permissions every time the device is connected.

In order to do this we need to intercept the USB_DEVICE_ATTACHED intent
action and check if Wolvic already has permissions to access the USB device.